Don't fail acquireToken when saving to cache fails

+++ ADAL/src/request/ADAuthenticationRequest.m
@@ -171,7 +171,7 @@ static dispatch_semaphore_t s_interactionLock = nil;
     {
         if (error)
         {
-            MSID_LOG_WARN(_requestParams, @"JSON desiarliazation error %ld", jsonError.code);
+            MSID_LOG_WARN(_requestParams, @"JSON desiarliazation error %ld", (long)jsonError.code);
             MSID_LOG_WARN_PII(_requestParams, @"JSON desiarliazation error %@ for claims %@", jsonError, claims);
 
             *error = [ADAuthenticationError errorFromAuthenticationError:AD_ERROR_DEVELOPER_INVALID_ARGUMENT
